+++ test/example_social.sh
@@ -45,7 +45,7 @@ cat <<EOF > $fexample
          key "member-id";
          description
            "List of members.";
-
+         ordered-by user;
          leaf member-id {
            type string {
              length "1..80";